clean and use buffers
