+++ Sofa/GL/src/sofa/gl/NewAxis.cpp
@@ -1,4 +1,4 @@
-/******************************************************************************
+﻿/******************************************************************************
 *                 SOFA, Simulation Open-Framework Architecture                *
 *                    (c) 2006 INRIA, USTL, UJF, CNRS, MGH                     *
 *                                                                             *
@@ -23,24 +23,316 @@
 
 #include <sofa/gl/gl.h>
 
-#include <cassert>
-#include <algorithm>
-#include <iostream>
+#include <sofa/type/Vec.h>
+#include <sofa/type/Mat.h>
+#include <sofa/type/Quat.h>
+#include <sofa/type/RGBAColor.h>
 
-constexpr int CYLINDER_SEGMENTS = 16;
+#include <unordered_map>
+#include <memory>
+#include <numbers>
+
+template <>
+struct std::hash<sofa::type::Vec3f>
+{
+    std::size_t operator()(const sofa::type::Vec3f& k) const
+    {
+        using std::size_t;
+        using std::hash;
+        using std::string;
+
+        return ((hash<float>()(k[0])
+            ^ (hash<float>()(k[1]) << 1)) >> 1)
+            ^ (hash<float>()(k[2]) << 1);
+    }
+};
+
+constexpr SReal PI = std::numbers::pi_v<SReal>;
+constexpr SReal PI2 = 2. * std::numbers::pi_v<SReal>;
+constexpr SReal PI_2 = std::numbers::pi_v<SReal> / 2.0;
+
+constexpr SReal constexpr_abs(SReal x) 
+{
+    return x < 0 ? -x : x;
+}
+
+
+constexpr SReal sin_fast(SReal x) 
+{
+    while (x > PI) x -= PI2;
+    while (x < -PI) x += PI2;
+
+    // Quick approximation using polynomial
+    const float B = 4.0f / PI;
+    const float C = -4.0f / (PI * PI);
+    const float P = 0.225f;
+
+    float y = B * x + C * x * constexpr_abs(x);
+    return P * (y * constexpr_abs(y) - y) + y;
+}
+
+constexpr SReal sin_cos(SReal x) 
+{
+    return sin_fast(x + PI_2);
+}
 
 namespace sofa::gl
 {
 
-std::map < type::Vec3f, CoordinateFrame > cacheCoordinateFrame;
+// Triangle structure
+struct Triangle {
+    int v0{}, v1{}, v2{};
+};
 
-// Function to calculate normal for a triangle
-type::Vec3 calculate_normal(const type::Vec3& v0, const type::Vec3& v1, const type::Vec3& v2) {
-    type::Vec3 edge1 = v1 - v0;
-    type::Vec3 edge2 = v2 - v0;
-    type::Vec3 normal = edge1.cross(edge2);
-    return normal.normalized();
-}
+// Cylinder mesh generator
+template<int Segments>
+struct CylinderMesh
+{
+    static constexpr int vertex_count = Segments * 2 + 2; // sides + 2 centers
+    static constexpr int triangle_count = Segments * 4;   // sides + caps
+
+    std::array<type::Vec3, vertex_count> vertices;
+    std::array<Triangle, triangle_count> triangles;
+    std::array<type::Vec3, triangle_count> normals;
+
+    constexpr CylinderMesh(const type::Vec3& start, const type::Vec3& end, double radius)
+    {
+        type::Vec3 direction = (end - start).normalized();
+
+        // Create perpendicular vectors
+        type::Vec3 up = type::Vec3(0, 1, 0);
+        if (constexpr_abs(direction.y()) > 0.9)
+        {
+            up = type::Vec3(1, 0, 0);
+        }
+
+        type::Vec3 right = up.cross(direction).normalized();
+        up = direction.cross(right).normalized();
+
+        // Generate vertices around the cylinder
+        int vertex_idx = 0;
+
+        // Side vertices (bottom and top circles)
+        for (int i = 0; i < Segments; ++i)
+        {
+            double angle = (static_cast<double>(i) / Segments) * 2.0 * std::numbers::pi_v<double>;
+            double cos_a = std::cos(angle);
+            double sin_a = std::sin(angle);
+
+            type::Vec3 offset = right * (cos_a * radius) + up * (sin_a * radius);
+
+            // Bottom circle
+            vertices[vertex_idx++] = start + offset;
+            // Top circle  
+            vertices[vertex_idx++] = end + offset;
+        }
+
+        //// Center points for caps
+        vertices[vertex_idx++] = start; // Bottom center
+        vertices[vertex_idx++] = end;   // Top center
+
+        // Generate triangles
+        int triangle_idx = 0;
+
+        // Side triangles
+        for (int i = 0; i < Segments; ++i)
+        {
+            int bottom_curr = i * 2;
+            int top_curr = i * 2 + 1;
+            int bottom_next = ((i + 1) % Segments) * 2;
+            int top_next = ((i + 1) % Segments) * 2 + 1;
+
+            // Two triangles per side segment
+            triangles[triangle_idx++] = Triangle(bottom_curr, top_curr, bottom_next);
+            triangles[triangle_idx++] = Triangle(top_curr, top_next, bottom_next);
+        }
+
+        // Cap triangles
+        int bottom_center = vertex_count - 2;
+        int top_center = vertex_count - 1;
+
+        for (int i = 0; i < Segments; ++i)
+        {
+            int bottom_curr = i * 2;
+            int top_curr = i * 2 + 1;
+            int bottom_next = ((i + 1) % Segments) * 2;
+            int top_next = ((i + 1) % Segments) * 2 + 1;
+
+            // Bottom cap (clockwise from below)
+            triangles[triangle_idx++] = Triangle(bottom_center, bottom_next, bottom_curr);
+            // Top cap (counter-clockwise from above)
+            triangles[triangle_idx++] = Triangle(top_center, top_curr, top_next);
+        }
+
+        // compute normals per vertex
+        for (int i = 0; i < triangle_count; ++i)
+        {
+            const auto& tri = triangles[i];
+            type::Vec3 v0 = vertices[tri.v0];
+            type::Vec3 v1 = vertices[tri.v1];
+            type::Vec3 v2 = vertices[tri.v2];
+            type::Vec3 normal = (v1 - v0).cross(v2 - v0) * -1;
+            normals[tri.v0] += normal;
+            normals[tri.v1] += normal;
+            normals[tri.v2] += normal;
+        }
+
+        //Not mandatory to do that if GL_NORMALIZE is set
+        for (int i = 0; i < vertex_count; ++i)
+        {
+            normals[i].normalize();
+        }
+    }
+};
+
+
+// Cone mesh generator
+template<int Segments>
+struct ConeMesh
+{
+    static constexpr int vertex_count = Segments + 2; // base circle + tip + base center
+    static constexpr int triangle_count = Segments * 2; // sides + base
+
+    std::array<type::Vec3, vertex_count> vertices;
+    std::array<Triangle, triangle_count> triangles;
+    std::array<type::Vec3, vertex_count> normals{};
+
+    constexpr ConeMesh(const type::Vec3& base, const type::Vec3& tip, double base_radius)
+    {
+        type::Vec3 direction = (tip - base).normalized();
+
+        // Create perpendicular vectors
+        type::Vec3 up = type::Vec3(0, 1, 0);
+        if (constexpr_abs(direction.y()) > 0.9) {
+            up = type::Vec3(1, 0, 0);
+        }
+
+        type::Vec3 right = up.cross(direction).normalized();
+        up = direction.cross(right).normalized();
+
+        // Generate vertices
+        int vertex_idx = 0;
+
+        // Tip vertex
+        vertices[vertex_idx++] = tip;
+
+        // Base circle vertices
+        for (int i = 0; i < Segments; ++i)
+        {
+            double angle = (static_cast<double>(i) / Segments) * 2.0 * std::numbers::pi_v<double>;
+            double cos_a = std::cos(angle);
+            double sin_a = std::sin(angle);
+
+            type::Vec3 offset = right * (cos_a * base_radius) + up * (sin_a * base_radius);
+            vertices[vertex_idx++] = base + offset;
+        }
+
+        // Base center
+        vertices[vertex_idx++] = base;
+
+        // Generate triangles
+        int triangle_idx = 0;
+        int tip_idx = 0;
+        int base_center_idx = vertex_count - 1;
+
+        // Side triangles
+        for (int i = 0; i < Segments; ++i)
+        {
+            int curr = i + 1;
+            int next = (i + 1) % Segments + 1;
+
+            // Side triangle (tip to base edge)
+            triangles[triangle_idx++] = Triangle(tip_idx, next, curr);
+
+            // Base triangle (base center to edge)
+            triangles[triangle_idx++] = Triangle(base_center_idx, curr, next);
+        }
+
+        // compute normals per vertex
+        for (int i = 0; i < triangle_count; ++i)
+        {
+            const auto& tri = triangles[i];
+            type::Vec3 v0 = vertices[tri.v0];
+            type::Vec3 v1 = vertices[tri.v1];
+            type::Vec3 v2 = vertices[tri.v2];
+            type::Vec3 normal = (v1 - v0).cross(v2 - v0) * -1;
+            normals[tri.v0] += normal;
+            normals[tri.v1] += normal;
+            normals[tri.v2] += normal;
+        }
+
+        //Not mandatory to do that if GL_NORMALIZE is set
+        for (int i = 0; i < vertex_count; ++i)
+        {
+            normals[i].normalize();
+        }
+    }
+};
+
+// Complete coordinate frame structure
+struct CoordinateFrame
+{
+    // Mesh types
+    inline static const int CYLINDER_SEGMENTS = 16;
+    using CylMesh = CylinderMesh<CYLINDER_SEGMENTS>;
+    using ConMesh = ConeMesh<CYLINDER_SEGMENTS>;
+
+    // Vertex and triangle counts
+    static constexpr int cylinder_vertex_count = CylMesh::vertex_count;
+    static constexpr int cylinder_triangle_count = CylMesh::triangle_count;
+    static constexpr int cone_vertex_count = ConMesh::vertex_count;
+    static constexpr int cone_triangle_count = ConMesh::triangle_count;
+
+    static constexpr int total_vertex_count = (cylinder_vertex_count + cone_vertex_count) * 3;
+    static constexpr int total_triangle_count = (cylinder_triangle_count + cone_triangle_count) * 3;
+
+    // Meshes for each axis
+    CylMesh x_cylinder;
+    ConMesh x_arrowhead;
+    CylMesh y_cylinder;
+    ConMesh y_arrowhead;
+    CylMesh z_cylinder;
+    ConMesh z_arrowhead;
+
+    struct FrameAxisParameters {
+        double axisLength{};
+        double arrowHeadLength{};
+        double cylinderRadius{};
+        double arrowHeadRadius{};
+    };
+
+    constexpr CoordinateFrame(const FrameAxisParameters& axisX, const FrameAxisParameters& axisY, const FrameAxisParameters& axisZ)
+        : x_cylinder(type::Vec3(0, 0, 0), type::Vec3(axisX.axisLength - axisX.arrowHeadLength, 0, 0), axisX.cylinderRadius)
+        , x_arrowhead(type::Vec3(axisX.axisLength - axisX.arrowHeadLength, 0, 0), type::Vec3(axisX.axisLength, 0, 0), axisX.arrowHeadRadius)
+        , y_cylinder(type::Vec3(0, 0, 0), type::Vec3(0, axisY.axisLength - axisY.arrowHeadLength, 0), axisY.cylinderRadius)
+        , y_arrowhead(type::Vec3(0, axisY.axisLength - axisY.arrowHeadLength, 0), type::Vec3(0, axisY.axisLength, 0), axisY.arrowHeadRadius)
+        , z_cylinder(type::Vec3(0, 0, 0), type::Vec3(0, 0, axisZ.axisLength - axisZ.arrowHeadLength), axisZ.cylinderRadius)
+        , z_arrowhead(type::Vec3(0, 0, axisZ.axisLength - axisZ.arrowHeadLength), type::Vec3(0, 0, axisZ.axisLength), axisZ.arrowHeadRadius)
+    {
+
+    }
+
+    // Get mesh data for specific axis components
+    struct MeshData {
+        const type::Vec3* vertices;
+        const Triangle* triangles;
+        const type::Vec3* normals;
+        int vertex_count;
+        int triangle_count;
+        int vertex_offset;
+    };
+
+    constexpr std::array<MeshData, 6> get_mesh_components() const {
+        return { {
+            {x_cylinder.vertices.data(), x_cylinder.triangles.data(), x_cylinder.normals.data(), cylinder_vertex_count, cylinder_triangle_count, 0},
+            {x_arrowhead.vertices.data(), x_arrowhead.triangles.data(), x_arrowhead.normals.data(), cone_vertex_count, cone_triangle_count, cylinder_vertex_count},
+            {y_cylinder.vertices.data(), y_cylinder.triangles.data(), y_cylinder.normals.data(), cylinder_vertex_count, cylinder_triangle_count, cylinder_vertex_count + cone_vertex_count},
+            {y_arrowhead.vertices.data(), y_arrowhead.triangles.data(), y_arrowhead.normals.data(), cone_vertex_count, cone_triangle_count, 2 * cylinder_vertex_count + cone_vertex_count},
+            {z_cylinder.vertices.data(), z_cylinder.triangles.data(), z_cylinder.normals.data(), cylinder_vertex_count, cylinder_triangle_count, 2 * cylinder_vertex_count + 2 * cone_vertex_count},
+            {z_arrowhead.vertices.data(), z_arrowhead.triangles.data(), z_arrowhead.normals.data(), cone_vertex_count, cone_triangle_count, 3 * cylinder_vertex_count + 2 * cone_vertex_count}
+        } };
+    }
+};
 
 // Render the complete coordinate frame
 void render_coordinate_frame(const CoordinateFrame& frame, const type::Vec3& center, const type::Quat<SReal>& orient, const type::Vec3& len, const type::RGBAColor& colorX, const type::RGBAColor& colorY, const type::RGBAColor& colorZ)
@@ -68,44 +360,31 @@ void render_coordinate_frame(const CoordinateFrame& frame, const type::Vec3& cen
 
     glMatrixMode(GL_MODELVIEW);
     glPushMatrix();
-    glTranslatef(center.x(), center.y(), center.z());
+    glTranslated(center.x(), center.y(), center.z());
     glRotated(phi * 180.0 / M_PI, 
         rotAxis.x(),
         rotAxis.y(),
         rotAxis.z());
-    
 
-    glBegin(GL_TRIANGLES);
-    // Render each component with its color
-    for (int i = 0; i < 6; ++i) 
+    glEnableClientState(GL_VERTEX_ARRAY);
+    glEnableClientState(GL_NORMAL_ARRAY);
+        
+    for (int i = 0; i < 6; ++i)
     {
         const auto& comp = mesh_components[i];
+        glColor4d(colors[i][0], colors[i][1], colors[i][2], colors[i][3]);
+        glVertexPointer(3, GL_DOUBLE, 0, comp.vertices);
+        glNormalPointer(GL_DOUBLE, 0, comp.normals);
+        glDrawElements(GL_TRIANGLES, comp.triangle_count * 3, GL_UNSIGNED_INT, comp.triangles);
 
-        glColor3f(colors[i][0], colors[i][1], colors[i][2]);
-
-        for (int i = 0; i < comp.triangle_count; ++i) 
-        {
-            const Triangle& tri = comp.triangles[i];
-            const type::Vec3& v0 = comp.vertices[tri.v0];
-            const type::Vec3& v1 = comp.vertices[tri.v1];
-            const type::Vec3& v2 = comp.vertices[tri.v2];
-
-            // Calculate and set normal for lighting
-            const type::Vec3 normal = calculate_normal(v0, v1, v2);
-            glNormal3f(normal.x(), normal.y(), normal.z());
-
-            // Emit vertices
-            glVertex3f(v0.x(), v0.y(), v0.z());
-            glVertex3f(v1.x(), v1.y(), v1.z());
-            glVertex3f(v2.x(), v2.y(), v2.z());
-        }
-
+        //glDrawArrays(GL_POINTS, 0, comp.vertex_count);
     }
-    glEnd();
 
     glPopMatrix();
     glPopAttrib();
 }
+
+std::unordered_map < type::Vec3f, CoordinateFrame > cacheCoordinateFrame;
 
 void Frame::draw(const type::Vec3& center, const Quaternion& orient, const type::Vec3& len, const type::RGBAColor& colorX, const type::RGBAColor& colorY, const type::RGBAColor& colorZ )
 {
@@ -131,9 +410,7 @@ void Frame::draw(const type::Vec3& center, const Quaternion& orient, const type:
         type::Vec3 lc(Lmax / 5_sreal, Lmax / 5_sreal, Lmax / 5_sreal); // = L / 5;
         type::Vec3 Lc = lc;
 
-        CoordinateFrame frame({ L[0], Lc[0], l[0], lc[0] }, { L[1], Lc[1], l[1], lc[1] }, { L[2], Lc[2], l[2], lc[2] });
-
-        cacheCoordinateFrame.emplace(len, frame);
+        cacheCoordinateFrame.emplace(len, CoordinateFrame ({ L[0], Lc[0], l[0], lc[0] }, { L[1], Lc[1], l[1], lc[1] }, { L[2], Lc[2], l[2], lc[2] }));
     }
 
     const auto& frame = cacheCoordinateFrame.at(len);
@@ -142,27 +419,32 @@ void Frame::draw(const type::Vec3& center, const Quaternion& orient, const type:
 
 void Frame::draw(const type::Vec3& center, const double orient[4][4], const type::Vec3& len, const type::RGBAColor& colorX, const type::RGBAColor& colorY, const type::RGBAColor& colorZ)
 {
-
-}
-
-void Frame::draw(const double *mat, const type::Vec3& len, const type::RGBAColor& colorX, const type::RGBAColor& colorY, const type::RGBAColor& colorZ)
-{
-
+    sofa::type::Matrix3 matOrient{ 
+        {orient[0][0], orient[0][1] , orient[0][2] },
+        {orient[1][0], orient[1][1] , orient[1][2] },
+        {orient[2][0], orient[2][1] , orient[2][2] } 
+    };
+    Quaternion q(sofa::type::QNOINIT);
+    q.fromMatrix(matOrient);
+    draw(center, q, len, colorX, colorY, colorZ);
 }
 
 void Frame::draw(const type::Vec3& center, const Quaternion& orient, SReal len, const type::RGBAColor& colorX, const type::RGBAColor& colorY, const type::RGBAColor& colorZ)
 {
-
+    draw(center, orient, { len, len, len }, colorX, colorY, colorZ);
 }
 
 void Frame::draw(const type::Vec3& center, const double orient[4][4], SReal len, const type::RGBAColor& colorX, const type::RGBAColor& colorY, const type::RGBAColor& colorZ)
 {
+    sofa::type::Matrix3 matOrient{
+        {orient[0][0], orient[0][1] , orient[0][2] },
+        {orient[1][0], orient[1][1] , orient[1][2] },
+        {orient[2][0], orient[2][1] , orient[2][2] }
+    };
 
-}
-
-void Frame::draw(const double *mat, SReal len, const type::RGBAColor& colorX, const type::RGBAColor& colorY, const type::RGBAColor& colorZ)
-{
-
+    Quaternion q(sofa::type::QNOINIT);
+    q.fromMatrix(matOrient);
+    draw(center, q, {len, len, len}, colorX, colorY, colorZ);
 }
 
 } // namespace sofa::gl
