+++ Sofa/GL/src/sofa/gl/NewAxis.h
@@ -22,336 +22,26 @@
 #pragma once
 #include <sofa/gl/config.h>
 
-#include <sofa/type/Vec.h>
-#include <sofa/type/Quat.h>
+#include <sofa/type/fwd.h>
 #include <sofa/type/RGBAColor.h>
 
-#include <sofa/topology/Triangle.h>
-
-#include <sofa/gl/gl.h>
-
-#include <map>
-#include <memory>
-#include <numbers>
 
 namespace sofa::gl
 {
 
-// Triangle structure
-struct Triangle {
-    int v0, v1, v2;
-    constexpr Triangle() : v0(0), v1(0), v2(0) {}
-    constexpr Triangle(int a, int b, int c) : v0(a), v1(b), v2(c) {}
-};
-
-namespace
-{
-    // Constexpr math helpers
-    constexpr SReal constexpr_abs(SReal x) {
-        return x < 0 ? -x : x;
-    }
-
-    constexpr SReal constexpr_cos(SReal x) {
-        // Taylor series approximation for cos(x)
-        // cos(x) = 1 - x²/2! + x⁴/4! - x⁶/6! + ...
-        SReal x2 = x * x;
-        SReal x4 = x2 * x2;
-        SReal x6 = x4 * x2;
-        SReal x8 = x6 * x2;
-        return 1.0 - x2 / 2.0 + x4 / 24.0 - x6 / 720.0 + x8 / 40320.0;
-    }
-
-    constexpr SReal constexpr_sin(SReal x) {
-        // Taylor series approximation for sin(x)
-        // sin(x) = x - x³/3! + x⁵/5! - x⁷/7! + ...
-        SReal x2 = x * x;
-        SReal x3 = x2 * x;
-        SReal x5 = x3 * x2;
-        SReal x7 = x5 * x2;
-        SReal x9 = x7 * x2;
-        return x - x3 / 6.0 + x5 / 120.0 - x7 / 5040.0 + x9 / 362880.0;
-    }
-}
-
-// Cylinder mesh generator
-template<int Segments>
-struct CylinderMesh {
-    static constexpr int vertex_count = Segments * 2 + 2; // sides + 2 centers
-    static constexpr int triangle_count = Segments * 4;   // sides + caps
-
-    std::array<type::Vec3, vertex_count> vertices;
-    std::array<Triangle, triangle_count> triangles;
-
-    constexpr CylinderMesh(const type::Vec3& start, const type::Vec3& end, double radius) {
-        type::Vec3 direction = (end - start).normalized();
-
-        // Create perpendicular vectors
-        type::Vec3 up = type::Vec3(0, 1, 0);
-        if (constexpr_abs(direction.y()) > 0.9) {
-            up = type::Vec3(1, 0, 0);
-        }
-
-        type::Vec3 right = up.cross(direction).normalized();
-        up = direction.cross(right).normalized();
-
-        // Generate vertices around the cylinder
-        int vertex_idx = 0;
-
-        // Side vertices (bottom and top circles)
-        for (int i = 0; i < Segments; ++i) {
-            double angle = (static_cast<double>(i) / Segments) * 2.0 * std::numbers::pi_v<double>;
-            double cos_a = std::cos(angle);
-            double sin_a = std::sin(angle);
-
-            type::Vec3 offset = right * (cos_a * radius) + up * (sin_a * radius);
-
-            // Bottom circle
-            vertices[vertex_idx++] = start + offset;
-            // Top circle  
-            vertices[vertex_idx++] = end + offset;
-        }
-
-        //// Center points for caps
-        vertices[vertex_idx++] = start; // Bottom center
-        vertices[vertex_idx++] = end;   // Top center
-
-        // Generate triangles
-        int triangle_idx = 0;
-
-        // Side triangles
-        for (int i = 0; i < Segments; ++i) {
-            int bottom_curr = i * 2;
-            int top_curr = i * 2 + 1;
-            int bottom_next = ((i + 1) % Segments) * 2;
-            int top_next = ((i + 1) % Segments) * 2 + 1;
-
-            // Two triangles per side segment
-            triangles[triangle_idx++] = Triangle(bottom_curr, top_curr, bottom_next);
-            triangles[triangle_idx++] = Triangle(top_curr, top_next, bottom_next);
-        }
-
-        // Cap triangles
-        int bottom_center = vertex_count - 2;
-        int top_center = vertex_count - 1;
-
-        for (int i = 0; i < Segments; ++i) {
-            int bottom_curr = i * 2;
-            int top_curr = i * 2 + 1;
-            int bottom_next = ((i + 1) % Segments) * 2;
-            int top_next = ((i + 1) % Segments) * 2 + 1;
-
-            // Bottom cap (clockwise from below)
-            triangles[triangle_idx++] = Triangle(bottom_center, bottom_next, bottom_curr);
-            // Top cap (counter-clockwise from above)
-            triangles[triangle_idx++] = Triangle(top_center, top_curr, top_next);
-        }
-    }
-};
-
-
-// Cone mesh generator
-template<int Segments>
-struct ConeMesh {
-    static constexpr int vertex_count = Segments + 2; // base circle + tip + base center
-    static constexpr int triangle_count = Segments * 2; // sides + base
-
-    std::array<type::Vec3, vertex_count> vertices;
-    std::array<Triangle, triangle_count> triangles;
-
-    constexpr ConeMesh(const type::Vec3& base, const type::Vec3& tip, double base_radius) {
-        type::Vec3 direction = (tip - base).normalized();
-
-        // Create perpendicular vectors
-        type::Vec3 up = type::Vec3(0, 1, 0);
-        if (constexpr_abs(direction.y()) > 0.9) {
-            up = type::Vec3(1, 0, 0);
-        }
-
-        type::Vec3 right = up.cross(direction).normalized();
-        up = direction.cross(right).normalized();
-
-        // Generate vertices
-        int vertex_idx = 0;
-
-        // Tip vertex
-        vertices[vertex_idx++] = tip;
-
-        // Base circle vertices
-        for (int i = 0; i < Segments; ++i) {
-            double angle = (static_cast<double>(i) / Segments) * 2.0 * std::numbers::pi_v<double>;
-            double cos_a = std::cos(angle);
-            double sin_a = std::sin(angle);
-
-            type::Vec3 offset = right * (cos_a * base_radius) + up * (sin_a * base_radius);
-            vertices[vertex_idx++] = base + offset;
-        }
-
-        // Base center
-        vertices[vertex_idx++] = base;
-
-        // Generate triangles
-        int triangle_idx = 0;
-        int tip_idx = 0;
-        int base_center_idx = vertex_count - 1;
-
-        // Side triangles
-        for (int i = 0; i < Segments; ++i) {
-            int curr = i + 1;
-            int next = (i + 1) % Segments + 1;
-
-            // Side triangle (tip to base edge)
-            triangles[triangle_idx++] = Triangle(tip_idx, next, curr);
-
-            // Base triangle (base center to edge)
-            triangles[triangle_idx++] = Triangle(base_center_idx, curr, next);
-        }
-    }
-};
-
-// Complete coordinate frame structure
-struct CoordinateFrame {
-    // Mesh types
-    inline static const int CYLINDER_SEGMENTS = 16;
-    using CylMesh = CylinderMesh<CYLINDER_SEGMENTS>;
-    using ConMesh = ConeMesh<CYLINDER_SEGMENTS>;
-
-    // Vertex and triangle counts
-    static constexpr int cylinder_vertex_count = CylMesh::vertex_count;
-    static constexpr int cylinder_triangle_count = CylMesh::triangle_count;
-    static constexpr int cone_vertex_count = ConMesh::vertex_count;
-    static constexpr int cone_triangle_count = ConMesh::triangle_count;
-
-    static constexpr int total_vertex_count = (cylinder_vertex_count + cone_vertex_count) * 3;
-    static constexpr int total_triangle_count = (cylinder_triangle_count + cone_triangle_count) * 3;
-
-    // Meshes for each axis
-    CylMesh x_cylinder;
-    ConMesh x_arrowhead;
-    CylMesh y_cylinder;
-    ConMesh y_arrowhead;
-    CylMesh z_cylinder;
-    ConMesh z_arrowhead;
-
-    struct FrameAxisParameters {
-        double axisLength{};
-        double arrowHeadLength{};
-        double cylinderRadius{};
-        double arrowHeadRadius{};
-    };
-
-    constexpr CoordinateFrame(const FrameAxisParameters& axisX, const FrameAxisParameters& axisY, const FrameAxisParameters& axisZ)
-        : x_cylinder(type::Vec3(0, 0, 0), type::Vec3(axisX.axisLength - axisX.arrowHeadLength, 0, 0), axisX.cylinderRadius)
-        , x_arrowhead(type::Vec3(axisX.axisLength - axisX.arrowHeadLength, 0, 0), type::Vec3(axisX.axisLength, 0, 0), axisX.arrowHeadRadius)
-        , y_cylinder(type::Vec3(0, 0, 0), type::Vec3(0, axisY.axisLength - axisY.arrowHeadLength, 0), axisY.cylinderRadius)
-        , y_arrowhead(type::Vec3(0, axisY.axisLength - axisY.arrowHeadLength, 0), type::Vec3(0, axisY.axisLength, 0), axisY.arrowHeadRadius)
-        , z_cylinder(type::Vec3(0, 0, 0), type::Vec3(0, 0, axisZ.axisLength - axisZ.arrowHeadLength), axisZ.cylinderRadius)
-        , z_arrowhead(type::Vec3(0, 0, axisZ.axisLength - axisZ.arrowHeadLength), type::Vec3(0, 0, axisZ.axisLength), axisZ.arrowHeadRadius)
-    {
-    }
-
-    // Get all vertices as a flat array
-    constexpr auto get_all_vertices() const {
-        std::array<type::Vec3, total_vertex_count> all_vertices{};
-        int index = 0;
-
-        // X-axis vertices
-        for (const auto& v : x_cylinder.vertices) all_vertices[index++] = v;
-        for (const auto& v : x_arrowhead.vertices) all_vertices[index++] = v;
-
-        // Y-axis vertices
-        for (const auto& v : y_cylinder.vertices) all_vertices[index++] = v;
-        for (const auto& v : y_arrowhead.vertices) all_vertices[index++] = v;
-
-        // Z-axis vertices
-        for (const auto& v : z_cylinder.vertices) all_vertices[index++] = v;
-        for (const auto& v : z_arrowhead.vertices) all_vertices[index++] = v;
-
-        return all_vertices;
-    }
-
-    // Get all triangles with proper vertex offsets
-    constexpr auto get_all_triangles() const {
-        std::array<Triangle, total_triangle_count> all_triangles{};
-        int index = 0;
-        int vertex_offset = 0;
-
-        // X-axis triangles
-        for (const auto& t : x_cylinder.triangles) {
-            all_triangles[index++] = Triangle(t.v0 + vertex_offset, t.v1 + vertex_offset, t.v2 + vertex_offset);
-        }
-        vertex_offset += cylinder_vertex_count;
-
-        for (const auto& t : x_arrowhead.triangles) {
-            all_triangles[index++] = Triangle(t.v0 + vertex_offset, t.v1 + vertex_offset, t.v2 + vertex_offset);
-        }
-        vertex_offset += cone_vertex_count;
-
-        // Y-axis triangles
-        for (const auto& t : y_cylinder.triangles) {
-            all_triangles[index++] = Triangle(t.v0 + vertex_offset, t.v1 + vertex_offset, t.v2 + vertex_offset);
-        }
-        vertex_offset += cylinder_vertex_count;
-
-        for (const auto& t : y_arrowhead.triangles) {
-            all_triangles[index++] = Triangle(t.v0 + vertex_offset, t.v1 + vertex_offset, t.v2 + vertex_offset);
-        }
-        vertex_offset += cone_vertex_count;
-
-        // Z-axis triangles
-        for (const auto& t : z_cylinder.triangles) {
-            all_triangles[index++] = Triangle(t.v0 + vertex_offset, t.v1 + vertex_offset, t.v2 + vertex_offset);
-        }
-        vertex_offset += cylinder_vertex_count;
-
-        for (const auto& t : z_arrowhead.triangles) {
-            all_triangles[index++] = Triangle(t.v0 + vertex_offset, t.v1 + vertex_offset, t.v2 + vertex_offset);
-        }
-
-        return all_triangles;
-    }
-
-    // Get mesh data for specific axis components
-    struct MeshData {
-        const type::Vec3* vertices;
-        const Triangle* triangles;
-        int vertex_count;
-        int triangle_count;
-        int vertex_offset;
-    };
-
-    constexpr std::array<MeshData, 6> get_mesh_components() const {
-        return { {
-            {x_cylinder.vertices.data(), x_cylinder.triangles.data(), cylinder_vertex_count, cylinder_triangle_count, 0},
-            {x_arrowhead.vertices.data(), x_arrowhead.triangles.data(), cone_vertex_count, cone_triangle_count, cylinder_vertex_count},
-            {y_cylinder.vertices.data(), y_cylinder.triangles.data(), cylinder_vertex_count, cylinder_triangle_count, cylinder_vertex_count + cone_vertex_count},
-            {y_arrowhead.vertices.data(), y_arrowhead.triangles.data(), cone_vertex_count, cone_triangle_count, 2 * cylinder_vertex_count + cone_vertex_count},
-            {z_cylinder.vertices.data(), z_cylinder.triangles.data(), cylinder_vertex_count, cylinder_triangle_count, 2 * cylinder_vertex_count + 2 * cone_vertex_count},
-            {z_arrowhead.vertices.data(), z_arrowhead.triangles.data(), cone_vertex_count, cone_triangle_count, 3 * cylinder_vertex_count + 2 * cone_vertex_count}
-        } };
-    }
-};
-
-
 class SOFA_GL_API Frame
 {
 public:
-    typedef sofa::type::Quat<SReal> Quaternion;
+    using Quaternion = sofa::type::Quat<SReal>;
 
-    Frame();
+    Frame() = delete;
     ~Frame() = delete;
 
     static void draw(const type::Vec3& center, const Quaternion& orient, const type::Vec3& length, const type::RGBAColor& colorX = type::RGBAColor::red(), const type::RGBAColor& colorY = type::RGBAColor::green(), const type::RGBAColor& colorZ = type::RGBAColor::red());
     static void draw(const type::Vec3& center, const double orient[4][4], const type::Vec3& length, const type::RGBAColor& colorX = type::RGBAColor::red(), const type::RGBAColor& colorY = type::RGBAColor::green(), const type::RGBAColor& colorZ = type::RGBAColor::red());
-    static void draw(const double* mat, const type::Vec3& length, const type::RGBAColor& colorX = type::RGBAColor::red(), const type::RGBAColor& colorY = type::RGBAColor::green(), const type::RGBAColor& colorZ = type::RGBAColor::red());
     static void draw(const type::Vec3& center, const Quaternion& orient, SReal length = 1.0_sreal, const type::RGBAColor& colorX = type::RGBAColor::red(), const type::RGBAColor& colorY = type::RGBAColor::green(), const type::RGBAColor& colorZ = type::RGBAColor::red());
     static void draw(const type::Vec3& center, const double orient[4][4], SReal length = 1.0_sreal, const type::RGBAColor& colorX = type::RGBAColor::red(), const type::RGBAColor& colorY = type::RGBAColor::green(), const type::RGBAColor& colorZ = type::RGBAColor::red());
-    static void draw(const double* mat, SReal length = 1.0_sreal, const type::RGBAColor& colorX = type::RGBAColor::red(), const type::RGBAColor& colorY = type::RGBAColor::green(), const type::RGBAColor& colorZ = type::RGBAColor::red());
-
-protected:
-    
-
-
-
+   
 };
 
 } // namespace sofa::gl
